# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  city?: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  /**
     * Email address or phone number
     * @minLength 1
     */
  identifier: string;
  /** @minLength 1 */
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  city?: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface ErrorResponse {
  message: string;
}

export interface UpdateProfileInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  city?: string;
}

export interface Address {
  id: number;
  userId: number;
  label: string;
  name: string;
  address: string;
  phone: string;
  pincode: string;
  city: string;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AddressInput {
  label: string;
  name: string;
  address: string;
  phone: string;
  pincode: string;
  city: string;
  isDefault: boolean;
}

export interface Booking {
  id: number;
  userId: number;
  bookingId: string;
  poojaId: string;
  poojaName: string;
  panditId: string;
  panditName: string;
  panditColor: string;
  panditInitials: string;
  date: string;
  time: string;
  amount: number;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface BookingInput {
  poojaId: string;
  poojaName: string;
  panditId: string;
  panditName: string;
  panditColor: string;
  panditInitials: string;
  date: string;
  time: string;
  amount: number;
}

export interface OrderItem {
  name: string;
  qty: number;
  price: number;
  unit: string;
}

export interface Order {
  id: number;
  userId: number;
  orderId: string;
  items: OrderItem[];
  amount: number;
  delivery: number;
  status: string;
  addressText: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  items: OrderItem[];
  amount: number;
  delivery: number;
  addressText: string;
}

export interface StatusUpdateInput {
  status: string;
}

export type PanditPoojasItem = {
  id: string;
  name: string;
  duration: string;
  price: number;
  includesPrasad: boolean;
};

export interface Pandit {
  id: number;
  name: string;
  shortName: string;
  specialty: string;
  category: string;
  rating: number;
  experience: string;
  bookings: number;
  age: number;
  city: string;
  address: string;
  available: string;
  specializations: string[];
  muhurats: string[];
  poojas: PanditPoojasItem[];
  initials: string;
  avatarColor: string;
  email: string;
  password: string;
  imageUrl?: string;
  createdAt: string;
  updatedAt: string;
}

export type PanditInputPoojasItem = {
  id: string;
  name: string;
  duration: string;
  price: number;
  includesPrasad: boolean;
};

export interface PanditInput {
  name: string;
  shortName: string;
  specialty: string;
  category: string;
  rating: number;
  experience: string;
  bookings: number;
  age: number;
  city: string;
  address: string;
  available: string;
  specializations: string[];
  muhurats: string[];
  poojas: PanditInputPoojasItem[];
  initials: string;
  avatarColor: string;
  email: string;
  password: string;
  imageUrl?: string;
}

export type DeleteAddress200 = {
  success: boolean;
};

export type PanditsForgotPasswordBody = {
  email: string;
};

export type PanditsForgotPassword200 = {
  success: boolean;
  resetLink: string;
};

export type DeletePandit200 = {
  success: boolean;
};