# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  label: string;
  name: string;
  address: string;
  phone: string;
  pincode: string;
  city: string;
  isDefault: boolean;
}