# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  poojaId: string;
  poojaName: string;
  panditId: string;
  panditName: string;
  panditColor: string;
  panditInitials: string;
  date: string;
  time: string;
  amount: number;
}