# lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  city?: string;
}