# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  name: string;
  email?: string;
  phone?: string;
  city?: string;
  createdAt: Date;
}